型ファイルの修正List.tsを削除し、Movie.tsにまとめる
